requirements.txt
discord.py[voice]==1.7.3
ffmpeg-python==0.2.0
yt-dlp==2022.7.18
